requirements.txt
attrs==19.3.0
backcall==0.2.0
backoff==1.10.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
iniconfig==1.0.1
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
kafka-python==2.0.1
mock==4.0.2
more-itertools==8.4.0
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.0.1
python-decouple==3.3
requests==2.24.0
six==1.15.0
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.10
validators==0.16.0
wcwidth==0.2.5
